+++ app/components/student_table_form.rb
@@ -29,8 +29,8 @@ class StudentTableForm < ViewComponent::Base
         </div>
       </div>
       <div class="table-cell flex justify-start gap-4 items-center">
-        <%= form.date_select :dob, { start_year: 1900, end_year: Date.today.year, selected: @student.dob || Date.new(Date.current.year, 1, 1) }, class: 'rounded-md border-purple-500 text-sm focus:border-green-600 focus:outline-none focus:ring-green-600' %>
-        <%= form.check_box :estimated_dob, checked: @student.estimated_dob , class: 'h-4 w-4 border-purple-500 text-green-600 focus:ring-green-600 cursor-pointer' %>
+        <%= render Datepicker.new(date: @student.dob || Date.new(Date.current.year, 1, 1), target: 'dob', form: form) %>
+        <%= form.check_box :estimated_dob, checked: @student.estimated_dob , class: 'h-4 w-4 border-purple-500 text-green-600 focus:ring-green-600 cursor-pointer ml-1' %>
         <label class="text-xs font-medium text-gray-700 cursor-pointer"><%= t(:dob_estimated) %></label>
         <%= render ValidationErrorComponent.new(model: @student, key: :dob) %>
       </div>