+++ app/javascript/application.js
@@ -3,5 +3,6 @@
 import "@hotwired/turbo-rails"
 import 'awesomplete'
 import 'controllers'
+
 // we don't want to enable turbo drive just yet until we finish the redesign
 // Turbo.session.drive = false;